+++ ui/api/resources/server.js
@@ -14,7 +14,11 @@ export default {
         return response.data;
     },
     stop: async () => {
-        const response = await client.get('/api/server/stop');
+        const response = await client.post('/api/server/stop');
+        return response.data;
+    },
+    restart: async () => {
+        const response = await client.post('/api/server/restart');
         return response.data;
     },
     start: async (ip, port, savefile) => {
@@ -26,7 +30,7 @@ export default {
         return response.data;
     },
     kill: async () => {
-        const response = await client.get('/api/server/kill');
+        const response = await client.post('/api/server/kill');
         return response.data;
     }
 }